Split omemo/utils.js into headless and UI versions
This replaces 'plugins/omemo' with 'plugins/omemo-views' and
'headless/plugins/omemo'. Most of the original code is still in
the former, but the contents of 'util' has been moved to the
headless version. Still pending: moving 'index.js' and all the
other files in the plugin!
