+++ src/plugins/omemo-views/device.js
