+++ src/plugins/omemo-views/devicelist.js
