+++ src/plugins/omemo-views/devices.js
