+++ src/plugins/omemo-views/templates/profile.js
@@ -1,5 +1,5 @@
 import spinner from "templates/spinner.js";
-import { formatFingerprint } from 'plugins/omemo/utils.js';
+import { formatFingerprint } from '@converse/headless/plugins/omemo/utils.js';
 import { html } from "lit";
 import { __ } from 'i18n';
